demo/heroku/requirements.txt
HTSQL
PyYAML
psycopg2
CherryPy
